♻️refactor: Redis 관련 구현체 / 인터페이스 패키지 변경 (#66)
