+++ src/main/java/kr/modernworld/modernworldv2/member/application/auth/listener/AuthEventListener.java
@@ -2,7 +2,7 @@ package kr.modernworld.modernworldv2.member.application.auth.listener;
 
 import kr.modernworld.modernworldv2.member.application.auth.event.LoginFailEvent;
 import kr.modernworld.modernworldv2.member.application.auth.event.LoginSuccessEvent;
-import kr.modernworld.modernworldv2.member.domain.token.RedisRepository;
+import kr.modernworld.modernworldv2.member.domain.auth.port.RedisRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.event.EventListener;
